🔨 refactor(LoraMesher.cpp, LoraMesher.h, MACService.cpp): rename WiFiService to MACService The WiFiService class has been renamed to MACService to better reflect its purpose of retrieving the MAC address of the device. The LoraMesher class has been updated to reflect this change. The MACService now uses the ESP32's chip MAC address instead of the WiFi MAC address to retrieve the local address.
🔥 chore(MACService.h, RoutingTableService.cpp, RoutingTableService.h, WifiService.cpp): replace WiFiService with MACService
The WiFiService class has been replaced with the MACService class. The MACService class provides a more accurate name for the functionality it provides. The changes were made to the MACService header file, RoutingTableService implementation file, and header file. The WiFiService implementation file was deleted as it is no longer needed.